==> analysis/05_Biodiversity.Rmd <==
---
title: "Biodiversity of male and female snail intestinal microbiomes across three temperature groups"
author: "Daphne Garcia"
date: "2025-04-10"
output: html_document
editor_options: 
  chunk_output_type: console
---


# Goals 

Our goal is to run an analysis on the within-sample biodiversity of the microbiome samples in this project. 

1. Calculate the Hill Numbers of the samples using the iNEXT package. 
2. Plot and evaluate the interpolated and extrapolated rarefaction curves from iNEXT. 
3. Analyze the Hill numbers as they relate to temperature and sex


Hypothesis:
My scientific question explores whether the temperature at which snails are raised, coupled with sex of the snail, has a significant effect on the richness of bacterial strains (ASVs) associated with them.

H0: There will be no difference in alpha diversity between the control, low-temperature, and high-temperature samples of male and female snails grouped separately

HA: There will be a significant difference in alpha diversity between the control, low-temperature, and high-temperature samples of male and female snails grouped separately. 


## Inputs 

1.  `phytree_preprocessed_physeq`, midpoint rooted tree 

## Outputs 

1. Calculated biodiversity measures (Hill numbers) across every sample. 
2. Figures visualizing the alpha diversity between different temperature-sex groups  
3. Statistical tests conveying the measured and quantified changes and patterns in biodiversity.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      fig.path = "figures/05_Biodiversity/")
    
```

# Set up the Environment 
```{r rmd-start}
# What time did we start running this script? 
start_time <- Sys.time()

set.seed(238428)

#install.packages("rstatix")
#install.packages("ggpubr")

pacman::p_load(tidyverse, devtools, patchwork, iNEXT, phyloseq,
               ggpubr, rstatix, install = FALSE)
source("code/colors.R")
```


## Load in Data 
```{r load-data}
load("data/04_PhylogeneticTree/phytree_preprocessed_physeqs.RData")

# take a look at it! 
midpoint_rooted_physeq

# Intuition check on seq depth
min(sample_sums(midpoint_rooted_physeq))

# Make a metadata dataframe 
metadata_df <- 
  midpoint_rooted_physeq %>%
  sample_data() %>%
  data.frame()

# view of it
glimpse(metadata_df)
```


# Biodiversity Hill Numbers with iNEXT

## Hill Number Equation
Runs iNEXT computation to produce rarefaction and extrapolation curves for species diversity, iNEXT data is loaded from `data/05_Biodiversity/iNEXT_data.RData`
```{r iNEXT-biodiversity}
# prepare the input data 
# Species in rows and the samples in columns 
iNEXT_input_df <-
  midpoint_rooted_physeq %>%
  otu_table() %>%
  data.frame()

# Inpect
dim(iNEXT_input_df)
iNEXT_input_df[1:5, 1:5]

# Run iNEXT 
# Remember to set the seed!
# ASVs in ROWs, samples in COLUMNS 
# iNEXT_data <- iNEXT(iNEXT_input_df, 
#                     q = c(0, 1, 2),
#                     datatype = "abundance")
# 
# save(iNEXT_data, file = "data/05_Biodiversity/iNEXT_data.RData")

load("data/05_Biodiversity/iNEXT_data.RData")
```

Since iNEXT takes a while to run, we will load in the data that was caluculated before! 
- took 1 hr 5 min

## Save iNEXT estimates to `data/05_Biodiversity/iNEXT_estimates_df.csv`
```{r load-iNEXT-data}

# Inspect 
str(iNEXT_data)
typeof(iNEXT_data)

inext_estimates_df <- iNEXT_data$iNextEst
 
# Dap- included: Write  the estimates to a csv file 
write.csv(inext_estimates_df, 
      file = "data/05_Biodiversity/iNEXT_estimates_df.csv")

```


# Rarefaction Curves 

## `ggiNEXT` Rarefaction Curves 
Create a colors dataframe to color the different experimental groups
```{r iNEXT-colors}
# Prepare Colors 
color_df <- 
  iNEXT_input_df %>%
  colnames() %>%
  data.frame()
# Check
head(color_df)
# Rename the column 
colnames(color_df)[1] <- "names"
# Check
head(color_df)


## fix metadata names to have 6 groups
#### Chatgpt: add snail_exp_group_name, for the six experimental groups
metadata_df <- metadata_df %>%
  mutate(snail_exp_group_name = substr(Sample.Name, 1, 2))


# Make a helper dataframe for plotting with colors 
iNEXT_color_df <- 
  color_df %>%
  # Fix the names for merging
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X",  replace = "")) %>%
  # Merge with metadata
  left_join(metadata_df, by = "names") %>%
  # Merge with colors for plotting with ggiNEXT    ###
  left_join(data.frame(station_colors = station_colors,
            snail_exp_group_name = names(station_colors)),
            by = "snail_exp_group_name")

head(iNEXT_color_df)
```

### Plotting rarefaction
Plot the rarefaction curves for richness, shannon, and simpson indices, colored by the sample's temperature-sex groups
```{r plot-iNEXT-rarefaction}
ggiNEXT(iNEXT_data, type = 1, facet.var = "Order.q") + 
  scale_color_manual(values = iNEXT_color_df$station_colors) + 
  scale_fill_manual(values = iNEXT_color_df$station_colors) + 
  scale_shape_manual(values = base::rep(17, nsamples(midpoint_rooted_physeq))) + 
  theme_bw() + 
  theme(legend.position = "none") + 
  labs(y = "Effective Number of ASVs", x = "Number of Sequences") 
```

### Manually plotting rarefaction
Plot the individual rarefaction curve for each experimental group (temperature-sex) for richness, shannon, and simpson indices
```{r manual-rarefaction}
str(iNEXT_data)

iNEXT_manual_df <- 
  iNEXT_data$iNextEst$size_based %>%
  dplyr::rename(names = Assemblage) %>%
  # fix the samples to merge with metadata 
  mutate(names = gsub(names, pattern = "[.]", replace = "-"),
         names = gsub(names, pattern = "X", replace = "")) %>%
  # join with the metadata
  left_join(., metadata_df, by = "names") %>%              #### Changed "station" to "snail_exp_group_name"
  mutate(snail_exp_group_name = factor(snail_exp_group_name, levels = c("CF",
                                              "LF",
                                              "HF",
                                              "CM",
                                              "LM",
                                              "HM")))

# Inspection
dim(iNEXT_manual_df)

# Manually plot rarefaction curves 
iNEXT_manual_df %>%
  # Filter out rows that are calculated with actual sequences (no extrapolated)
  dplyr::filter(Method == "Rarefaction") %>%
  # Now, let's draw the plot! 
  ggplot(aes(x = m, y = qD, color = snail_exp_group_name, group = names)) +  #Changed "station" to "snail_exp_group_name"
  geom_line() + 
  # Facet by station 
  facet_grid(Order.q~snail_exp_group_name, scales = "free") +               #Changed "station" to "snail_exp_group_name"
  scale_color_manual(values = station_colors) + 
  theme_bw() + 
  labs(y = "Effective Number of ASVs", x = "Number of Sequences") + 
  theme(legend.position = "bottom")
```

INTERPRETATION #1: What can you conclude from the ggiNEXT and manual rarefaction curve plots? Are there “enough” sequences to analyze the samples? Is it valid to make any conclusions about species richness when q = 0? What about when q = 1? Or q = 2? Why or why not?

I think it is safe to say that for q = 1 and q = 2, which emphasize the common and dominant species, enough sequences have been sampled. In all 6 experimental groups (control, high-temperature, low-temperature vs female, male), all rarefaction curves plateau. 

For richness, the low and high temperature female groups (LF and HF) definitely don't plateau. There are several samples in these two groups that do not asymptote. in FLF and HF, there is even a slight bimodal distribution. Because these values don't plateau, we can't say anything about the richness.


For shannon, the trends are more amplified in HF, where the samples that didn't asymptote have extremely high number of ASVs that are moderately abundant. HF and CM (control-temperature male) have an outlier. 

For simpson, every sample pleateaus, however in LF, HF, CM, and HM, there is one outlier that has many more dominant ASVs than the rest. This, coupled with the outliers in shannon and richness, may mean that these samples have a very even abundance, causing all ASV's to be considered "rare", "normal" and "dominant" at the same time. 


### Run Shapiro-Wilklins
Test if the data is normalized using a Shapiro-Wilklins test
```{r shapiro-wilk-richness}

### Now what about for richness? 
obs_div_df <- 
  iNEXT_manual_df %>%
  dplyr::filter(Method == "Observed") 

# check it
glimpse(obs_div_df)

# Pull out unique data from the "Control-temperature Female" group
obs_rich_df <- 
  obs_div_df %>%
  dplyr::filter(Order.q == 0)

# Test of the data is normal for the continuous value of richness
shapiro.test(obs_rich_df$qD)
# 
# ## Make a for loop to print shaprio wilkins of all 6 experimental groups
# for (group in c("CF", "LF", "HF", "CM", "LM", "HM")) {
#   obs_group_rich_df <- 
#     obs_div_df %>%
#     dplyr::filter(snail_exp_group_name == group) %>%
#     dplyr::filter(Order.q == 0)
#   print(c("Shapiro test for group: ", group))
#   print(shapiro.test(obs_group_rich_df$qD))
# }
```


NOTES:
W ≈ 1 → data is close to normal.
W < 1 → increasing deviation from normality.

Null hypothesis (H₀): The data are normally distributed.
p-value > 0.05: Fail to reject H₀ and the data appears normal.
Alternative hypothesis (H₁): The data are not normally distributed.
p-value < 0.05: We can reject the H₀, which will indicate that the data is not normal.

MY DATA: 

When I break up the samples by their group (CF LF HF CM LM HM), all W's are above 0.9, and p values > 0.05, except for HF
in the high-temperature female group, W = 0.82, p = 0.02, which means the data is not normal. 
However, when all of them are grouped together, W = 0.71931, p-value < 2.2e-16. The p value is extreeemely low which means we have to reject the null hypothesis, and the data is definitely not normalized.



# Diversity by snail experimental group

```{r make-obs-df}
# Plot boxplots by station against diversity
obs_div_df %>%
  ggplot(aes(x = snail_exp_group_name, y = qD, fill = snail_exp_group_name, 
             color = snail_exp_group_name)) + 
  facet_wrap(.~Order.q, scales = "free_y") + 
  geom_jitter() + 
  geom_boxplot(alpha = 0.5, outlier.shape = NA) + 
  scale_color_manual(values = station_colors) + 
  scale_fill_manual(values = station_colors) + 
  labs(y = "Effective Number of ASVs") + 
  theme_bw() + 
  theme(legend.position = "bottom",
        axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 30,
                                   hjust = 1, 
                                   vjust = 1))
```


### Statistically test Richness
Statistically test ASV richness across snail experimental groups using a Kruskal-Wallis test and Dunns post-hoc
```{r station-diversity-richness, fig.height=4, fig.width=4}
# Pull out richness data 
obs_rich_df <- 
  obs_div_df %>%
  dplyr::filter(Order.q == 0)

# Calculate the kruskall-wallis stats
kw_group_rich <- 
  kruskal_test(qD ~ snail_exp_group_name, data = obs_rich_df) 

# Look at it 
kw_group_rich

# post-hoc test with Dunn's Test
dunn_group_rich <- 
  obs_rich_df %>%
  dunn_test(qD ~ snail_exp_group_name, p.adjust.method = "fdr") %>%
  add_xy_position(x = "snail_exp_group_name")

# look at the pairwise
dunn_group_rich

```


## Plot hill-richness 
Create a boxplot of hill-richness across snail experimental groups. include Kruskal-wallis test
```{r richness-diversity, fig.height=4, fig.width=4}

# Richness by station 
rich_ggboxplot <- 
  ggboxplot(obs_rich_df, x = "snail_exp_group_name", y = "qD", 
          color = "snail_exp_group_name", fill = "snail_exp_group_name", alpha = 0.3,
          outlier.shape = NA) + 
  # Add points
  geom_jitter(aes(color = snail_exp_group_name)) + 
  scale_color_manual(values = station_colors) + 
  scale_fill_manual(values = station_colors) + 
  labs(y = "# of Total ASVs (Richness)") + 
  # Add kruskal & Tukey
  stat_compare_means(method = "kruskal.test", label.y = min(obs_rich_df$qD)) +  # KRUSKAL p-value
  geom_pwc(
    aes(group = snail_exp_group_name), tip.length = 0, hide.ns = TRUE,
    method = "wilcox.test", label = "p.adj.format",
    p.adjust.method = "fdr", bracket.nudge.y = -0.08) + 
  theme_bw() + 
  theme(legend.position = "none",
        axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 30,
                                   hjust = 1, 
                                   vjust = 1))

rich_ggboxplot

```


## Statistically test and plot Shannon diversity
Statistically test Shannon index across snail experimental groups using a Kruskal-Wallis test and Dunns post-hoc. Create a boxplot of hill-shannon across snail experimental groups.
```{r shannon-diversity, fig.height=4, fig.width=4}
# Pull out shann data 
obs_shan_df <- 
  obs_div_df %>%
  dplyr::filter(Order.q == 1)

# Calculate the kruskall wallis since the data NOT NORMAL
kw_group_shan <- 
   kruskal_test(qD ~ snail_exp_group_name, data = obs_shan_df)
# Look at it 
kw_group_shan

# post-hoc test with Dunn's Test
dunn_group_shan <- 
  obs_shan_df %>%
  dunn_test(qD ~ snail_exp_group_name, p.adjust.method = "fdr") %>%
  add_xy_position(x = "snail_exp_group_name")
# look at the pairwise
dunn_group_shan


# Shannon by experimental group
shan_ggboxplot <- 
  ggboxplot(obs_shan_df, x = "snail_exp_group_name", y = "qD", 
          color = "snail_exp_group_name", fill = "snail_exp_group_name", alpha = 0.3,
          outlier.shape = NA) + 
  # Add points
  geom_jitter(aes(color = snail_exp_group_name)) + 
  scale_color_manual(values = station_colors) + 
  scale_fill_manual(values = station_colors) + 
  labs(y = "# of Dominant ASVs (Shannon)") +
  # Add kruskal and wilcoxon
  
  stat_compare_means(method = "kruskal.test", label.y = 80) +  # KRUSKAL p-value (not normal)
  geom_pwc(
    aes(group = snail_exp_group_name), tip.length = 0, hide.ns = TRUE,
    method = "wilcox.test", label = "p.adj.signif",
    step.increase = 0.08,
    p.adjust.method = "fdr", bracket.nudge.y = -0.08) + 
  theme_bw() + 
  theme(legend.position = "none",
        axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 30,
                                   hjust = 1, 
                                   vjust = 1))

shan_ggboxplot

```

## Statistically test and plot Simpson diversity
Statistically test Simpson index across snail experimental groups using a Kruskal-Wallis test and Dunns post-hoc. Create a boxplot of hill-simpson across snail experimental groups.
```{r simpson-diversity, fig.height=4, fig.width=4}
# Pull out simpson data 
obs_simps_df <- 
  obs_div_df %>%
  dplyr::filter(Order.q == 2)

# Calculate the kruskall wallis since the data NOT NORMAL
kw_group_simps <- 
   kruskal_test(qD ~ snail_exp_group_name, data = obs_simps_df)
# Look at it 
kw_group_simps

# post-hoc test with Dunn's Test
dunn_group_simps <- 
  obs_simps_df %>%
  dunn_test(qD ~ snail_exp_group_name, p.adjust.method = "fdr") %>%
  add_xy_position(x = "snail_exp_group_name")
# look at the pairwise
dunn_group_simps


# Simpson by experimental group
simps_ggboxplot <- 
  ggboxplot(obs_simps_df, x = "snail_exp_group_name", y = "qD", 
          color = "snail_exp_group_name", fill = "snail_exp_group_name", alpha = 0.3,
          outlier.shape = NA) + 
  # Add points
  geom_jitter(aes(color = snail_exp_group_name)) + 
  scale_color_manual(values = station_colors) + 
  scale_fill_manual(values = station_colors) + 
  labs(y = "# of Dominant ASVs (Simpson)") +
  # Add kruskal and wilcoxon
  
  stat_compare_means(method = "kruskal.test", label.y = 80) +  # KRUSKAL p-value (not normal)
  geom_pwc(
    aes(group = snail_exp_group_name), tip.length = 0, hide.ns = TRUE,
    method = "wilcox.test", label = "p.adj.signif",
    step.increase = 0.08,
    p.adjust.method = "fdr", bracket.nudge.y = -0.08) + 
  theme_bw() + 
  theme(legend.position = "none",
        axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 30,
                                   hjust = 1, 
                                   vjust = 1))
simps_ggboxplot

```

### Plot all diversities together
```{r Plot_hill_diversities}

rich_ggboxplot + shan_ggboxplot + simps_ggboxplot
```


INTERPRETATION #2: If you have categorical data, what can you conclude regarding the role that all ASVs (richess, q = 0) versus common ASVs (Shannon, q = 1) versus dominant ASVs (Simpson, q = 2) have on the biodiversity of your samples as it relates to your variable of interest?


Richness: When comparing all to all, every experimental group had statistically the same richness to one another. This means that temperature has no effect on the number of rare ASVs in a snail's gut, which proves our null hypothesis.
Shannon: in all experimental groups, except for control-temperature female and high-temperature male had the same shannon index. This further confirms the null hypothesis suggesting that temperature has no effect on the number of normally abundant ASVs in snail gut. 
Simpson: Here, we start seeing statistical differences between the experimental groups. CF is significantly different to HF and HM, additionally LM is significantly different to HM, which is exciting! There are statistically significant differences between low-temperature and high-temperature experimental groups. This suggests that for abundant taxons, high temperature results in a significant increase of dominant taxons. This correlates with a finding in the paper, where increased temperatures allowed opportunistic pathogens to proliferate.


INTERPRETATION #3: How did you statistically test your scientific question in Interpretation #2? Why?

I tested my question with a Kruskall-Wallis test with Dunn post-hoc. The reason is that my data did not follow a normal distribution: When I ran the shapiro wilkins test W = 0.71931, p-value < 2.2e-16. The p value is extremely low (2.2e-16) which rejected the null hypothesis of the data being normalized. Additionally  I had 6 experimental groups, low-temperature, high-temperature, and control-temperature for female and male host snails, which is why I had to use KW instead of a Wilcoxon rank-sum test. I added a Dunn post-hoc to determine which specific groups differ from each other. 


# For the presentation
1. Add the extra variables (temperature and sex)
2. do shapiro-wilkins for richness, shannon, simpson just for temp
3. Make the three plots for just temperature
```{r for_presentation}

### Now what about for richness? 
# obs_div_df <- 
#   iNEXT_manual_df %>%
#   dplyr::filter(Method == "Observed") %>%
#   mutate(temperature = substr(Sample.Name, 1, 1)) %>%
#   mutate(sex = substr(Sample.Name, 2, 2))
# 
# ########################################################################################
# 
# obs_rich_df <- 
#   obs_div_df %>%
#   dplyr::filter(Order.q == 0)
# 
# # Test of the data is normal for the continuous value of richness
# shapiro.test(obs_rich_df$qD)
# 
# # post-hoc test with Dunn's Test
# dunn_group_rich <- 
#   obs_rich_df %>%
#   dunn_test(qD ~ snail_exp_group_name, p.adjust.method = "fdr") %>%
#   add_xy_position(x = "snail_exp_group_name")
# 
# # look at the pairwise
# dunn_group_rich
# 
# 
# # Richness by station 
# rich_ggboxplot <- 
#   ggboxplot(obs_rich_df, x = "temperature", y = "qD", 
#           color = "temperature", fill = "temperature", alpha = 0.3,
#           outlier.shape = NA) + 
#   # Add points
#   geom_jitter(aes(color = temperature)) + 
#   scale_color_manual(values = temperature_colors) + 
#   scale_fill_manual(values = temperature_colors) + 
#   labs(y = "# of Total ASVs (Richness)") + 
#   # Add kruskal & Tukey
#   stat_compare_means(method = "kruskal.test", label.y = min(obs_rich_df$qD)) +  # KRUSKAL p-value
#   geom_pwc(
#     aes(group = snail_exp_group_name), tip.length = 0, hide.ns = TRUE,
#     method = "wilcox.test", label = "p.adj.format",
#     p.adjust.method = "fdr", bracket.nudge.y = -0.08) + 
#   theme_bw() + 
#   theme(legend.position = "none",
#         axis.title.x = element_blank(),
#         axis.text.x = element_text(angle = 30,
#                                    hjust = 1, 
#                                    vjust = 1))
# 
# rich_ggboxplot
# 
# ###########################################################################
# 
# obs_shan_df <- 
#   obs_div_df %>%
#   dplyr::filter(Order.q == 1)
# 
# # test if shannon data is normal 
# shapiro.test(obs_shan_df$qD)
#   #not normal at all
# # Calculate the kruskall wallis since the data NOT NORMAL
# 
# 
# 
# # Shannon by experimental group
# shan_ggboxplot <- 
#   ggboxplot(obs_shan_df, x = "temperature", y = "qD", 
#           color = "temperature", fill = "temperature", alpha = 0.3,
#           outlier.shape = NA) + 
#   # Add points
#   geom_jitter(aes(color = temperature)) + 
#   scale_color_manual(values = temperature_colors) + 
#   scale_fill_manual(values = temperature_colors) + 
#   labs(y = "# of Dominant ASVs (Shannon)") +
#   # Add kruskal and wilcoxon
#   
#   stat_compare_means(method = "kruskal.test", label.y = 5) +  # KRUSKAL p-value (not normal)
#   geom_pwc(
#     aes(group = snail_exp_group_name), tip.length = 0, hide.ns = TRUE,
#     method = "wilcox.test", label = "p.adj.signif",
#     step.increase = 0.08,
#     p.adjust.method = "fdr", bracket.nudge.y = -0.08) + 
#   theme_bw() + 
#   theme(legend.position = "none",
#         axis.title.x = element_blank(),
#         axis.text.x = element_text(angle = 30,
#                                    hjust = 1, 
#                                    vjust = 1))
# 
# shan_ggboxplot
# 
# #############################################################################
# 
# # Pull out simpson data 
# obs_simps_df <- 
#   obs_div_df %>%
#   dplyr::filter(Order.q == 2)
# 
# # test if simpson data is normal 
# shapiro.test(obs_simps_df$qD)
#   # not at all
# 
# 
# # Simpson by experimental group
# simps_ggboxplot <- 
#   ggboxplot(obs_simps_df, x = "temperature", y = "qD", 
#           color = "temperature", fill = "temperature", alpha = 0.3,
#           outlier.shape = NA) + 
#   # Add points
#   geom_jitter(aes(color = temperature)) + 
#   scale_color_manual(values = temperature_colors) + 
#   scale_fill_manual(values = temperature_colors) + 
#   labs(y = "# of Dominant ASVs (Simpson)") +
#   # Add kruskal and wilcoxon
#   
#   stat_compare_means(method = "kruskal.test", label.y = 1) +  # KRUSKAL p-value (not normal)
#   geom_pwc(
#     aes(group = snail_exp_group_name), tip.length = 0, hide.ns = TRUE,
#     method = "wilcox.test", label = "p.adj.signif",
#     step.increase = 0.08,
#     p.adjust.method = "fdr", bracket.nudge.y = -0.08) + 
#   theme_bw() + 
#   theme(legend.title = element_text(face="bold"),
#         axis.title.x = element_blank(),
#         axis.text.x = element_text(angle = 30,
#                                    hjust = 1, 
#                                    vjust = 1))
# 
# + theme(legend.title = element_text(colour="blue", size=10, 
#                                       face="bold"))
# 
# simps_ggboxplot
# 
# 
# #########################################################
# 
# rich_ggboxplot + shan_ggboxplot + simps_ggboxplot

```


```{r end}
# Take the time now that we are at the end of the script
end_time <- Sys.time()
end_time 
```

devtools
```{r devtools}
devtools::session_info()
```
